refactor: use impl trait in examples
