+++ examples/common/mod.rs
@@ -1,0 +1,62 @@
+use std::sync::Arc;
+
+use async_trait::async_trait;
+use futures::{stream, StreamExt};
+
+use pgwire::api::auth::noop::NoopStartupHandler;
+use pgwire::api::query::SimpleQueryHandler;
+use pgwire::api::results::{DataRowEncoder, FieldFormat, FieldInfo, QueryResponse, Response, Tag};
+use pgwire::api::{ClientInfo, Type};
+use pgwire::error::PgWireResult;
+
+pub struct DummyProcessor;
+
+impl NoopStartupHandler for DummyProcessor {}
+
+#[async_trait]
+impl SimpleQueryHandler for DummyProcessor {
+    async fn do_query<'a, C>(&self, _client: &mut C, query: &str) -> PgWireResult<Vec<Response<'a>>>
+    where
+        C: ClientInfo + Unpin + Send + Sync,
+    {
+        println!("{:?}", query);
+        if query.starts_with("SELECT") {
+            let f1 = FieldInfo::new("id".into(), None, None, Type::INT4, FieldFormat::Text);
+            let f2 = FieldInfo::new("name".into(), None, None, Type::VARCHAR, FieldFormat::Text);
+            let schema = Arc::new(vec![f1, f2]);
+
+            let data = vec![
+                (Some(0), Some("Tom")),
+                (Some(1), Some("Jerry")),
+                (Some(2), None),
+            ];
+            let schema_ref = schema.clone();
+            let data_row_stream = stream::iter(data.into_iter()).map(move |r| {
+                let mut encoder = DataRowEncoder::new(schema_ref.clone());
+                encoder.encode_field(&r.0)?;
+                encoder.encode_field(&r.1)?;
+
+                encoder.finish()
+            });
+
+            Ok(vec![Response::Query(QueryResponse::new(
+                schema,
+                data_row_stream,
+            ))])
+        } else {
+            Ok(vec![Response::Execution(Tag::new("OK").with_rows(1))])
+        }
+    }
+}
+
+pub struct DummyProcessorFactory {
+    pub handler: Arc<DummyProcessor>,
+}
+
+impl DummyProcessorFactory {
+    pub fn new() -> DummyProcessorFactory {
+        Self {
+            handler: Arc::new(DummyProcessor),
+        }
+    }
+}
