+++ examples/scram.rs
@@ -4,7 +4,6 @@ use std::sync::Arc;
 
 use async_trait::async_trait;
 
-use pgwire::api::cancel::NoopCancelHandler;
 use rustls_pemfile::{certs, pkcs8_private_keys};
 use rustls_pki_types::{CertificateDer, PrivateKeyDer};
 use tokio::net::TcpListener;
@@ -12,30 +11,12 @@ use tokio_rustls::rustls::ServerConfig;
 use tokio_rustls::TlsAcceptor;
 
 use pgwire::api::auth::scram::{gen_salted_password, SASLScramAuthStartupHandler};
-use pgwire::api::auth::{AuthSource, DefaultServerParameterProvider, LoginInfo, Password};
-use pgwire::api::copy::NoopCopyHandler;
-use pgwire::api::query::{PlaceholderExtendedQueryHandler, SimpleQueryHandler};
-use pgwire::api::results::{Response, Tag};
-
-use pgwire::api::{ClientInfo, NoopErrorHandler, PgWireServerHandlers};
+use pgwire::api::auth::{
+    AuthSource, DefaultServerParameterProvider, LoginInfo, Password, StartupHandler,
+};
+use pgwire::api::PgWireServerHandlers;
 use pgwire::error::PgWireResult;
 use pgwire::tokio::process_socket;
-
-pub struct DummyProcessor;
-
-#[async_trait]
-impl SimpleQueryHandler for DummyProcessor {
-    async fn do_query<'a, C>(
-        &self,
-        _client: &mut C,
-        _query: &str,
-    ) -> PgWireResult<Vec<Response<'a>>>
-    where
-        C: ClientInfo + Unpin + Send + Sync,
-    {
-        Ok(vec![Response::Execution(Tag::new("OK").with_rows(1))])
-    }
-}
 
 pub fn random_salt() -> Vec<u8> {
     Vec::from(rand::random::<[u8; 10]>())
@@ -75,27 +56,11 @@ fn setup_tls() -> Result<TlsAcceptor, IOError> {
 }
 
 struct DummyProcessorFactory {
-    handler: Arc<DummyProcessor>,
     cert: Vec<u8>,
 }
 
 impl PgWireServerHandlers for DummyProcessorFactory {
-    type StartupHandler = SASLScramAuthStartupHandler<DummyAuthDB, DefaultServerParameterProvider>;
-    type SimpleQueryHandler = DummyProcessor;
-    type ExtendedQueryHandler = PlaceholderExtendedQueryHandler;
-    type CopyHandler = NoopCopyHandler;
-    type CancelHandler = NoopCancelHandler;
-    type ErrorHandler = NoopErrorHandler;
-
-    fn simple_query_handler(&self) -> Arc<Self::SimpleQueryHandler> {
-        self.handler.clone()
-    }
-
-    fn extended_query_handler(&self) -> Arc<Self::ExtendedQueryHandler> {
-        Arc::new(PlaceholderExtendedQueryHandler)
-    }
-
-    fn startup_handler(&self) -> Arc<Self::StartupHandler> {
+    fn startup_handler(&self) -> Arc<impl StartupHandler> {
         let mut authenticator = SASLScramAuthStartupHandler::new(
             Arc::new(DummyAuthDB),
             Arc::new(DefaultServerParameterProvider::default()),
@@ -107,27 +72,12 @@ impl PgWireServerHandlers for DummyProcessorFactory {
 
         Arc::new(authenticator)
     }
-
-    fn copy_handler(&self) -> Arc<Self::CopyHandler> {
-        Arc::new(NoopCopyHandler)
-    }
-
-    fn cancel_handler(&self) -> Arc<Self::CancelHandler> {
-        Arc::new(NoopCancelHandler)
-    }
-
-    fn error_handler(&self) -> Arc<Self::ErrorHandler> {
-        Arc::new(NoopErrorHandler)
-    }
 }
 
 #[tokio::main]
 pub async fn main() {
     let cert = fs::read("examples/ssl/server.crt").unwrap();
-    let factory = Arc::new(DummyProcessorFactory {
-        handler: Arc::new(DummyProcessor),
-        cert,
-    });
+    let factory = Arc::new(DummyProcessorFactory { cert });
 
     let server_addr = "127.0.0.1:5432";
     let tls_acceptor = setup_tls().unwrap();
