+++ examples/server.rs
@@ -1,136 +1,27 @@
-use std::fmt::Debug;
 use std::sync::Arc;
 
-use async_trait::async_trait;
-use futures::{stream, Sink, SinkExt, StreamExt};
-use pgwire::api::cancel::NoopCancelHandler;
 use tokio::net::TcpListener;
 
-use pgwire::api::auth::noop::NoopStartupHandler;
-use pgwire::api::copy::NoopCopyHandler;
-use pgwire::api::query::{PlaceholderExtendedQueryHandler, SimpleQueryHandler};
-use pgwire::api::results::{DataRowEncoder, FieldFormat, FieldInfo, QueryResponse, Response, Tag};
-use pgwire::api::{ClientInfo, NoopErrorHandler, PgWireServerHandlers, Type};
-use pgwire::error::ErrorInfo;
-use pgwire::error::{PgWireError, PgWireResult};
-use pgwire::messages::response::NoticeResponse;
-use pgwire::messages::{PgWireBackendMessage, PgWireFrontendMessage};
+use pgwire::api::auth::StartupHandler;
+use pgwire::api::query::SimpleQueryHandler;
+use pgwire::api::PgWireServerHandlers;
 use pgwire::tokio::process_socket;
 
-pub struct DummyProcessor;
+mod common;
 
-#[async_trait]
-impl NoopStartupHandler for DummyProcessor {
-    async fn post_startup<C>(
-        &self,
-        client: &mut C,
-        _message: PgWireFrontendMessage,
-    ) -> PgWireResult<()>
-    where
-        C: ClientInfo + Sink<PgWireBackendMessage> + Unpin + Send,
-        C::Error: Debug,
-        PgWireError: From<<C as Sink<PgWireBackendMessage>>::Error>,
-    {
-        println!(
-            "connected {:?}: {:?}",
-            client.socket_addr(),
-            client.metadata()
-        );
-        Ok(())
-    }
-}
-
-#[async_trait]
-impl SimpleQueryHandler for DummyProcessor {
-    async fn do_query<'a, C>(&self, client: &mut C, query: &str) -> PgWireResult<Vec<Response<'a>>>
-    where
-        C: ClientInfo + Sink<PgWireBackendMessage> + Unpin + Send + Sync,
-        C::Error: Debug,
-        PgWireError: From<<C as Sink<PgWireBackendMessage>>::Error>,
-    {
-        client
-            .send(PgWireBackendMessage::NoticeResponse(NoticeResponse::from(
-                ErrorInfo::new(
-                    "NOTICE".to_owned(),
-                    "01000".to_owned(),
-                    format!("Query received {}", query),
-                ),
-            )))
-            .await?;
-
-        if query.starts_with("SELECT") {
-            let f1 = FieldInfo::new("id".into(), None, None, Type::INT4, FieldFormat::Text);
-            let f2 = FieldInfo::new("name".into(), None, None, Type::VARCHAR, FieldFormat::Text);
-            let schema = Arc::new(vec![f1, f2]);
-
-            let data = vec![
-                (Some(0), Some("Tom")),
-                (Some(1), Some("Jerry")),
-                (Some(2), None),
-            ];
-            let schema_ref = schema.clone();
-            let data_row_stream = stream::iter(data.into_iter()).map(move |r| {
-                let mut encoder = DataRowEncoder::new(schema_ref.clone());
-                encoder.encode_field(&r.0)?;
-                encoder.encode_field(&r.1)?;
-
-                encoder.finish()
-            });
-
-            // tokio::time::sleep(Duration::from_millis(10000)).await;
-
-            Ok(vec![Response::Query(QueryResponse::new(
-                schema,
-                data_row_stream,
-            ))])
-        } else {
-            Ok(vec![Response::Execution(Tag::new("OK").with_rows(1))])
-        }
-    }
-}
-
-struct DummyProcessorFactory {
-    handler: Arc<DummyProcessor>,
-}
-
-impl PgWireServerHandlers for DummyProcessorFactory {
-    type StartupHandler = DummyProcessor;
-    type SimpleQueryHandler = DummyProcessor;
-    type ExtendedQueryHandler = PlaceholderExtendedQueryHandler;
-    type CopyHandler = NoopCopyHandler;
-    type ErrorHandler = NoopErrorHandler;
-    type CancelHandler = NoopCancelHandler;
-
-    fn simple_query_handler(&self) -> Arc<Self::SimpleQueryHandler> {
+impl PgWireServerHandlers for common::DummyProcessorFactory {
+    fn simple_query_handler(&self) -> Arc<impl SimpleQueryHandler> {
         self.handler.clone()
     }
 
-    fn extended_query_handler(&self) -> Arc<Self::ExtendedQueryHandler> {
-        Arc::new(PlaceholderExtendedQueryHandler)
-    }
-
-    fn startup_handler(&self) -> Arc<Self::StartupHandler> {
+    fn startup_handler(&self) -> Arc<impl StartupHandler> {
         self.handler.clone()
-    }
-
-    fn copy_handler(&self) -> Arc<Self::CopyHandler> {
-        Arc::new(NoopCopyHandler)
-    }
-
-    fn cancel_handler(&self) -> Arc<Self::CancelHandler> {
-        Arc::new(NoopCancelHandler)
-    }
-
-    fn error_handler(&self) -> Arc<Self::ErrorHandler> {
-        Arc::new(NoopErrorHandler)
     }
 }
 
 #[tokio::main]
 pub async fn main() {
-    let factory = Arc::new(DummyProcessorFactory {
-        handler: Arc::new(DummyProcessor),
-    });
+    let factory = Arc::new(common::DummyProcessorFactory::new());
 
     let server_addr = "127.0.0.1:5432";
     let listener = TcpListener::bind(server_addr).await.unwrap();
